refactor(engine): extract daemon state layer to src/daemon/state/
SessionState, TerminalSignal, StuckConfig/StuckSignal, and all pure state
functions extracted from workflow-runner.ts to src/daemon/state/. Three
focused modules: session-state.ts (state + transitions), terminal-signal.ts
(first-writer-wins gate), stuck-detection.ts (pure evaluation function).

Introduces 5 named state transitions that replace direct field writes:
- advanceStep(): atomic push+count+token update (was 3 inline writes)
- recordCompletion(): atomic isComplete+notes+artifacts (was 3 inline writes)
- updateToken(): names the blocked-retry token update intent
- setSessionId(): names the async workrailSessionId resolve
- recordToolCall(): ring-buffer push+shift enforced in one place

Architecture test updated to scan state/terminal-signal.ts for the one
allowed state.terminalSignal assignment (inside setTerminalSignal()).
workflow-runner.ts must have zero direct assignments -- test enforces this.

Removed re-exports from workflow-runner.ts: TerminalSignal, SessionState,
StuckConfig, StuckSignal, setTerminalSignal, evaluateStuckSignals,
createSessionState (all re-exported from state/ for backward compat).

377/377 tests pass. Build clean.
